Move stepper settings to "Advanced Settings" menu (#13063)
